Refactor for Testing and Docbloc added
To aid tests which need to be added I have refactored the whole project to allow
the service class to be injected instead of extended. This has not changed any
general useage of the module only the internal workings.

Docblocs have been added to all code for ease of understanding.